Adding support for up to 16.000 segments per level.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11190 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/gpgpu_lighting.h
@@ -46,7 +46,7 @@ class CGPGPULighting {
 		void Begin (void);
 		void End (void);
 		void InitShader (void);
-		int Compute (short nVertex, int nState, CFaceColor *colorP);
+		int Compute (int nVertex, int nState, CFaceColor *colorP);
 
 	private:
 		GLuint CreateBuffer (int i);
